Declarative library
